WoT restructure D1a: render/table.ts -> render/tableRenderer.ts
